create interface for adding/removing appearances

+++ TheSongList/Controllers/AppearancesController.cs
@@ -1,11 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using TheSongList.Models.Entities;
 using TheSongList.Services;
 
@@ -20,39 +19,12 @@ namespace TheSongList.Controllers
             _context = context;
         }
 
-        // GET: Appearances
-        public async Task<IActionResult> Index()
-        {
-            var songContext = _context.Appearances.Include(a => a.Episode).Include(a => a.Song);
-            return View(await songContext.ToListAsync());
-        }
-
-        // GET: Appearances/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var appearance = await _context.Appearances
-                .Include(a => a.Episode)
-                .Include(a => a.Song)
-                .FirstOrDefaultAsync(m => m.EpisodeId == id);
-            if (appearance == null)
-            {
-                return NotFound();
-            }
-
-            return View(appearance);
-        }
-
         // GET: Appearances/Create
         [Authorize(Policy = "CanEdit")]
         public IActionResult Create()
         {
-            ViewData["EpisodeId"] = new SelectList(_context.Episodes, "Id", "Id");
-            ViewData["SongId"] = new SelectList(_context.Songs, "Id", "Name");
+            ViewData["EpisodeId"] = new SelectList(GetEpisodes(), "Id", "Name");
+            ViewData["SongId"] = new SelectList(GetSongs(), "Id", "Name");
             return View();
         }
 
@@ -66,85 +38,52 @@ namespace TheSongList.Controllers
         {
             if (ModelState.IsValid)
             {
-                _context.Add(appearance);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["EpisodeId"] = new SelectList(_context.Episodes, "Id", "Id", appearance.EpisodeId);
-            ViewData["SongId"] = new SelectList(_context.Songs, "Id", "Name", appearance.SongId);
-            return View(appearance);
-        }
-
-        // GET: Appearances/Edit/5
-        [Authorize(Policy = "CanEdit")]
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var appearance = await _context.Appearances.FindAsync(id);
-            if (appearance == null)
-            {
-                return NotFound();
-            }
-            ViewData["EpisodeId"] = new SelectList(_context.Episodes, "Id", "Id", appearance.EpisodeId);
-            ViewData["SongId"] = new SelectList(_context.Songs, "Id", "Name", appearance.SongId);
-            return View(appearance);
-        }
-
-        // POST: Appearances/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [Authorize(Policy = "CanEdit")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("EpisodeId,SongId")] Appearance appearance)
-        {
-            if (id != appearance.EpisodeId)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
+                if (!await AppearanceExists(appearance.SongId, appearance.EpisodeId))
                 {
-                    _context.Update(appearance);
+                    _context.Add(appearance);
                     await _context.SaveChangesAsync();
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!AppearanceExists(appearance.EpisodeId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Details", "Songs", new { id = appearance.SongId });
             }
-            ViewData["EpisodeId"] = new SelectList(_context.Episodes, "Id", "Id", appearance.EpisodeId);
-            ViewData["SongId"] = new SelectList(_context.Songs, "Id", "Name", appearance.SongId);
+            ViewData["EpisodeId"] = new SelectList(GetEpisodes(), "Id", "Name", appearance.EpisodeId);
+            ViewData["SongId"] = new SelectList(GetSongs(), "Id", "Name", appearance.SongId);
             return View(appearance);
         }
 
+        private IEnumerable<object> GetEpisodes() => _context.Episodes
+            .Include(e => e.Season)
+            .OrderBy(e => e.Season.SortOrder)
+            .ThenBy(e => e.EpisodeNumber)
+            .Select(e => new
+            {
+                e.Id,
+                Name = $"S{e.Season.Name} E{e.EpisodeNumber}" + (!string.IsNullOrEmpty(e.Name) ? $" - {e.Name}" : string.Empty)
+            });
+
+        private IEnumerable<object> GetSongs() => _context.Songs
+            .Include(s => s.Artist)
+            .OrderBy(s => s.Name)
+            .Select(s => new
+            {
+                s.Id,
+                Name = s.Name + (!string.IsNullOrEmpty(s.Artist.Name) ? $" - {s.Artist.Name}" : string.Empty)
+            });
+
         // GET: Appearances/Delete/5
+        [HttpGet("Appearances/Delete/{episodeId}/{songId}")]
         [Authorize(Policy = "CanEdit")]
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete([FromRoute]int? episodeId, [FromRoute]int? songId)
         {
-            if (id == null)
+            if (!episodeId.HasValue || !songId.HasValue)
             {
                 return NotFound();
             }
 
             var appearance = await _context.Appearances
                 .Include(a => a.Episode)
+                .ThenInclude(e => e.Season)
                 .Include(a => a.Song)
-                .FirstOrDefaultAsync(m => m.EpisodeId == id);
+                .FirstOrDefaultAsync(m => m.EpisodeId == episodeId && m.SongId == songId);
             if (appearance == null)
             {
                 return NotFound();
@@ -154,20 +93,21 @@ namespace TheSongList.Controllers
         }
 
         // POST: Appearances/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost("Appearances/Delete/{episodeId}/{songId}"), ActionName("Delete")]
         [Authorize(Policy = "CanEdit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed([FromRoute]int episodeId, [FromRoute]int songId)
         {
-            var appearance = await _context.Appearances.FindAsync(id);
-            _context.Appearances.Remove(appearance);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var appearance = await _context.Appearances.FirstOrDefaultAsync(a => a.EpisodeId == episodeId && a.SongId == songId);
+            if (appearance != null)
+            {
+                _context.Appearances.Remove(appearance);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction("Details", "Songs", new { id = appearance.SongId });
         }
 
-        private bool AppearanceExists(int id)
-        {
-            return _context.Appearances.Any(e => e.EpisodeId == id);
-        }
+        private async Task<bool> AppearanceExists(int songId, int episodeId) =>
+            await _context.Appearances.AnyAsync(e => e.EpisodeId == episodeId && e.SongId == songId);
     }
 }
